[FLINK-21151] Make RocksDBFullSnapshotResources a toplevel class
